kimvieware-orchestrator/uploads/job_0029_blog-back-main/blog-back-main/articles/permissions.py
from rest_framework.permissions import BasePermission, SAFE_METHODS

class IsAdminOrReadOnly(BasePermission):
    """
    Lecture pour tous, écriture uniquement pour les admins
    """
    def has_permission(self, request, view):
        if request.method in SAFE_METHODS:
            return True
        return request.user and request.user.is_authenticated and request.user.is_staff
