kimvieware-orchestrator/uploads/job_0029_blog-back-main/blog-back-main/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.0
Django==5.2.11
django-cors-headers==4.3.1
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0

